requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
PyPDF2
speechrecognition
pyttsx3
pyaudio